+++ src/test/java/tests/TestsSetUp.java
@@ -1,0 +1,4 @@
+package tests;
+
+public class TestsSetUp {
+}
